+++ services/database/tests/unit/test_db_cli_extended.py
@@ -80,3 +80,32 @@ class TestDBCliExtended(unittest.TestCase):
         assert 'version' not in result.output
         assert 'Error: Missing option(s)' in result.output
         assert '--vm-cluster-id' in result.output
+
+    def test_patch_database(self):
+        # Test by not providing mandatory args
+        result = util.invoke_command(['db', 'database', 'patch'])
+        assert 'Error: Missing option(s)' in result.output
+        assert '--database-id' in result.output
+        assert '--patch-id' not in result.output
+        assert '--patch-action' not in result.output
+        assert '--one-off-patches' not in result.output
+
+        # Test by providing both RU and one-off patch, but missing database-id
+        result = util.invoke_command(['db', 'database', 'patch', '--one-off-patches', 'oneOff', '--patch-id', 'patchId'])
+        assert 'Error: Missing option(s)' in result.output
+        assert '--database-id' in result.output
+
+        # Test by not providing both RU and one-off patch, which shall fail as none to patch provided
+        result = util.invoke_command(['db', 'database', 'patch', '--database-id', 'databaseId'])
+        assert 'Specify either \'--one-off-patches\' or \'--patch-id and --patch-action\'. Requesting both is not supported.' in result.output
+
+        result = util.invoke_command(['db', 'database', 'patch', '--database-id', 'databaseId', '--one-off-patches', 'oneOff', '--patch-id', 'patchId'])
+        assert 'Specify either \'--one-off-patches\' or \'--patch-id and --patch-action\'. Requesting both is not supported.' in result.output
+        result = util.invoke_command(['db', 'database', 'patch', '--database-id', 'databaseId', '--one-off-patches', 'oneOff', '--patch-action', 'patchAction'])
+        assert 'Specify either \'--one-off-patches\' or \'--patch-id and --patch-action\'. Requesting both is not supported.' in result.output
+        result = util.invoke_command(['db', 'database', 'patch', '--database-id', 'databaseId', '--one-off-patches', 'oneOff', '--patch-id', 'patchId', '--patch-action', 'patchAction'])
+        assert 'Specify either \'--one-off-patches\' or \'--patch-id and --patch-action\'. Requesting both is not supported.' in result.output
+        result = util.invoke_command(['db', 'database', 'patch', '--database-id', 'databaseId', '--patch-action', 'patchAction'])
+        assert 'Specify either \'--one-off-patches\' or \'--patch-id and --patch-action\'. Requesting both is not supported.' in result.output
+        result = util.invoke_command(['db', 'database', 'patch', '--database-id', 'databaseId', '--patch-id', 'patchId'])
+        assert 'Specify either \'--one-off-patches\' or \'--patch-id and --patch-action\'. Requesting both is not supported.' in result.output